LayoutTests/http/tests/webrtc/enumerateDevicesInFrames-expected.txt

CONSOLE MESSAGE: line 18: Feature policy 'Camera' check failed for iframe with origin 'http://127.0.0.1:8000' and allow attribute 'microphone:'none'; camera:'none''.
CONSOLE MESSAGE: line 18: Feature policy 'Microphone' check failed for iframe with origin 'http://127.0.0.1:8000' and allow attribute 'microphone:'none'; camera:'none''.
CONSOLE MESSAGE: line 18: Not allowed to call enumerateDevices.


PASS allow attribute can be updated dynamically 
